python/ScrapingBee/requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
python-dotenv==1.0.0
requests==2.31.0
scrapingbee==1.2.0
urllib3==2.0.4
